refactor: consolidate discussion forum into app/discussion/
Seventh phase — merges discussion-related code from three different
places into one Django-style app:
- 7 discussion/post/like API routes, previously a block inside
  app/blueprints/courses.py
- course_discussions/discussion_detail page renders, previously in
  app/blueprints/main.py (their templates renamed into
  app/discussion/templates/discussion/)
- the old standalone app/blueprints/discussion.py (55 lines: edit/delete
  post), now folded in

Found and resolved a genuine pre-existing bug while merging: both
courses.py's `delete_post` and discussion.py's `api_delete_post` were
registered for the identical route (DELETE /api/posts/<id>), since both
blueprints used url_prefix='/api'. Empirically verified with the
unmodified app (`url_map.bind().match()`) that Werkzeug was always
dispatching to courses.py's version — discussion.py's copy (which
additionally checked verify_course_in_school and restricted deletion to
the post author or the *teacher* specifically, not just the discussion
creator) was dead, unreachable code. Kept the actually-active
implementation as the single canonical `discussion.delete_post`; the
shadowed copy is gone. Flagging this for visibility: the surviving
implementation does not call get_school_id_or_abort/verify_course_in_school,
which the dead copy did — worth a follow-up security review, but fixing
the check itself is a behavior change outside this refactor's scope.

Page routes needed a second, prefix-free `discussion_pages_bp` in the
same module (mirroring the issues_pages_bp pattern from Phase 4), since
discussion_bp's url_prefix='/api' can't also serve `/kelas/<id>/diskusi`.

app/models/discussion.py -> app/discussion/models.py, fixing its
`from . import db` (only worked inside the old app/models/ package) to
`from app.core.extensions import db`. tests/test_discussion.py moved to
app/discussion/tests/ (self-contained, no shared fixtures needed beyond
what's already in app/conftest.py).

Verified: create_app() now reports 223 routes (224 minus the one dead
duplicate rule removed), SQLAlchemy metadata still has 33 tables, and
pytest is unchanged from baseline (25 pre-existing failures, 99
passing) with test_discussion.py fully green in its new location.
