+++ app/blueprints/courses.py
@@ -6,7 +6,7 @@ from flask import Blueprint, request, jsonify, abort, current_app
 from flask_login import login_required, current_user
 from werkzeug.utils import secure_filename
 from sqlalchemy.orm import joinedload, selectinload
-from app.models import db, Course, AcademicYear, UserRole, Link, File, Discussion, Post, Like, UserCourseOrder, Quiz, GradeType, QuizStatus, ContentFolder, Assignment
+from app.models import db, Course, AcademicYear, UserRole, Link, File, UserCourseOrder, Quiz, GradeType, QuizStatus, ContentFolder, Assignment
 from app.helpers import sanitize_text, sanitize_rich_text, is_valid_color, is_valid_class_code, generate_class_code, get_courses_for_user, format_course_data, log_activity
 from app.core.authorization import get_school_id_or_abort, verify_course_in_school, verify_academic_year_in_school
 
@@ -420,166 +420,6 @@ def api_create_file(course_id):
         }), 201
 
     return jsonify({'success': False, 'message': 'Terjadi kesalahan saat mengunggah file'}), 500
-
-
-@courses_bp.route('/courses/<int:course_id>/discussions', methods=['POST'])
-@login_required
-def create_discussion(course_id):
-    course = db.session.get(Course, course_id)
-    if not course:
-        return jsonify({'success': False, 'message': 'Mata pelajaran tidak ditemukan'}), 404
-
-    school_id = get_school_id_or_abort()
-    verify_course_in_school(course, school_id)
-
-    if current_user.role != UserRole.GURU and current_user not in course.students:
-        return jsonify({'success': False, 'message': 'Anda tidak memiliki izin untuk membuat diskusi di kelas ini'}), 403
-
-    data = request.get_json() or {}
-    title = sanitize_text(data.get('title', ''), max_len=200)
-    content = sanitize_text(data.get('content', ''))
-
-    if not title or not content:
-        return jsonify({'success': False, 'message': 'Judul dan isi diskusi wajib diisi'}), 400
-
-    try:
-        new_discussion = Discussion(
-            title=title,
-            course_id=course_id,
-            user_id=current_user.id
-        )
-        db.session.add(new_discussion)
-        db.session.commit()
-
-        new_post = Post(
-            content=content,
-            discussion_id=new_discussion.id,
-            user_id=current_user.id
-        )
-        db.session.add(new_post)
-        db.session.commit()
-
-        return jsonify({'success': True, 'discussion': {'id': new_discussion.id, 'title': new_discussion.title}}), 201
-    except Exception as e:
-        db.session.rollback()
-        logger.error(f"Failed to create discussion for course {course_id}: {e}", exc_info=True)
-        return jsonify({'success': False, 'message': 'Terjadi kesalahan saat membuat diskusi'}), 500
-
-
-@courses_bp.route('/courses/<int:course_id>/discussions', methods=['GET'])
-@login_required
-def get_discussions(course_id):
-    course = db.session.get(Course, course_id)
-    if not course:
-        return jsonify({'success': False, 'message': 'Mata pelajaran tidak ditemukan'}), 404
-
-    school_id = get_school_id_or_abort()
-    verify_course_in_school(course, school_id)
-
-    if current_user.role != UserRole.GURU and current_user not in course.students:
-        return jsonify({'success': False, 'message': 'Anda tidak memiliki izin untuk melihat diskusi di kelas ini'}), 403
-
-    discussions = Discussion.query.filter_by(course_id=course_id)\
-        .options(joinedload(Discussion.user), selectinload(Discussion.posts).joinedload(Post.user))\
-        .order_by(Discussion.created_at.desc()).all()
-    return jsonify({'success': True, 'discussions': [d.to_dict() for d in discussions]})
-
-
-@courses_bp.route('/discussions/<int:discussion_id>/posts', methods=['GET'])
-@login_required
-def get_posts(discussion_id):
-    discussion = db.session.get(Discussion, discussion_id)
-    if not discussion:
-        return jsonify({'success': False, 'message': 'Diskusi tidak ditemukan'}), 404
-
-    if current_user.role != UserRole.GURU and current_user not in discussion.course.students:
-        return jsonify({'success': False, 'message': 'Anda tidak memiliki izin untuk melihat diskusi ini'}), 403
-
-    posts = Post.query.filter_by(discussion_id=discussion_id)\
-        .options(joinedload(Post.user), selectinload(Post.replies).joinedload(Post.user), selectinload(Post.likes).joinedload(Like.user))\
-        .order_by(Post.created_at.asc()).all()
-    return jsonify({'success': True, 'posts': [p.to_dict() for p in posts]})
-
-
-@courses_bp.route('/discussions/<int:discussion_id>/posts', methods=['POST'])
-@login_required
-def add_post(discussion_id):
-    discussion = db.session.get(Discussion, discussion_id)
-    if not discussion:
-        return jsonify({'success': False, 'message': 'Diskusi tidak ditemukan'}), 404
-
-    if discussion.closed:
-        return jsonify({'success': False, 'message': 'Diskusi ini sudah ditutup'}), 403
-
-    data = request.get_json() or {}
-    content = sanitize_text(data.get('content', ''))
-    parent_id = data.get('parent_id')
-
-    if not content:
-        return jsonify({'success': False, 'message': 'Isi respon tidak boleh kosong'}), 400
-
-    new_post = Post(
-        content=content,
-        discussion_id=discussion_id,
-        user_id=current_user.id,
-        parent_id=parent_id
-    )
-    db.session.add(new_post)
-    db.session.commit()
-
-    return jsonify({'success': True, 'post': new_post.to_dict()}), 201
-
-
-@courses_bp.route('/posts/<int:post_id>/like', methods=['POST'])
-@login_required
-def like_post(post_id):
-    post = db.session.get(Post, post_id)
-    if not post:
-        return jsonify({'success': False, 'message': 'Post tidak ditemukan'}), 404
-
-    like = Like.query.filter_by(post_id=post_id, user_id=current_user.id).first()
-    if like:
-        db.session.delete(like)
-        db.session.commit()
-        return jsonify({'success': True, 'action': 'unliked'})
-    else:
-        new_like = Like(post_id=post_id, user_id=current_user.id)
-        db.session.add(new_like)
-        db.session.commit()
-        return jsonify({'success': True, 'action': 'liked'})
-
-
-@courses_bp.route('/posts/<int:post_id>', methods=['DELETE'])
-@login_required
-def delete_post(post_id):
-    post = db.session.get(Post, post_id)
-    if not post:
-        return jsonify({'success': False, 'message': 'Post tidak ditemukan'}), 404
-
-    # Allow deletion if user is the post author or the discussion creator (teacher)
-    if current_user.id != post.user_id and current_user.id != post.discussion.user_id:
-        return jsonify({'success': False, 'message': 'Anda tidak memiliki izin untuk menghapus post ini'}), 403
-
-    db.session.delete(post)
-    db.session.commit()
-
-    return jsonify({'success': True})
-
-
-@courses_bp.route('/discussions/<int:discussion_id>/close', methods=['POST'])
-@login_required
-def close_discussion(discussion_id):
-    discussion = db.session.get(Discussion, discussion_id)
-    if not discussion:
-        return jsonify({'success': False, 'message': 'Diskusi tidak ditemukan'}), 404
-
-    if discussion.user_id != current_user.id:
-        return jsonify({'success': False, 'message': 'Hanya pembuat diskusi yang dapat menutupnya'}), 403
-
-    discussion.closed = True
-    db.session.commit()
-
-    return jsonify({'success': True, 'message': 'Diskusi telah ditutup'})
 
 
 @courses_bp.route('/courses/reorder', methods=['POST'])
