+++ app/discussion/models.py
@@ -1,7 +1,7 @@
 import datetime
 from typing import List, Optional
 from sqlalchemy.orm import Mapped, mapped_column, relationship
-from . import db
+from app.core.extensions import db
 from app.helpers import get_jakarta_now
 
 class Discussion(db.Model):
