+++ app/discussion/templates/discussion/course_discussions.html
